added starvation, death messages
also made some minor changes to the portrait drawing. definitely needs more work though. i think im going to need to split up the setupGame() function between a function called at the very start of the game and one for the start of each specific run. also an opacity variable for the portrait drawing function.

yeah overall this is more of a minor patch than a full on update, expect another update pretty soon (but not soo soon i got college to focus on)