cleanup/bug fixes
cleaned up some code, made it so it will actually notify you when you are starving (within 10 seconds of death by hunger), fixed the bug where occasionally youd get stuck in a wall after going through a door. havent worked on this for a while but i have some new ideas for what id like to add to this game. we'll see if i have the time for it.